[MRG] Remove boston dataset following sklearn deprecation, replace by leukemia (#33)
* Remove boston dataset following sklearn deprecation, replace by leukemia

* add generic libsvm dataset

* rm leukemia file

* Update datasets/libsvm.py


* leukemia back

* add comment

* Update datasets/leukemia.py
